# docs/swagger.yaml
definitions:
...
  controller.GetUserResp:
    properties:
      avatar:
        type: string
      campus:
        type: string
      hobby:
        type: string
      id:
        type: integer
      money:
        type: integer
      more:
        type: string
      nickName:
        type: string
      openId:
        type: string
      phone:
        type: string
      setting1:
        type: integer
      setting2:
        type: integer
      setting3:
        type: integer
      sex:
        type: integer
      trueName:
        type: string
    type: object
  controller.PersonalPage:
    properties:
      Songs:
        items:
          $ref: '#/definitions/models.Songs'
        type: array
      admire:
        items:
          $ref: '#/definitions/models.Admire'
        type: array
      avatar:
        type: string
      background:
        type: string
      hide_number:
        type: integer
      more:
        type: string
      name:
        type: string
      phone:
        type: string
      requestSongs:
        items:
          $ref: '#/definitions/models.RequestSongs'
        type: array
      school:
        type: string
      setting1:
        type: integer
      setting2:
        type: integer
      setting3:
        type: integer
      truename:
        type: string
    type: object
...
  controller.PutUserInf:
    properties:
      avatar:
        type: string
      name:
        type: string
      phone:
        type: string
      setting1:
        type: integer
      setting2:
        type: integer
      setting3:
        type: integer
      signature:
        type: string
      truename:
        type: string
    type: object
...
  controller.UserRegister:
    properties:
      name:
        type: string
      phone:
        type: string
      realname:
        type: string
      school:
        type: string
      sex:
        type: integer
    type: object
...
  e.ErrMsgResponse:
    properties:
      message:
        type: string
    type: object
  models.Admire:
    properties:
      from:
        type: string
      id:
        type: integer
      number:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
  models.AllRank:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Rank'
        type: array
      time:
        type: string
    type: object
...
  models.Rank:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      photo:
        type: string
      praise:
        type: integer
      source:
        type: string
      text:
        type: string
      time:
        type: string
      type:
        type: integer
      user:
        type: string
    type: object
  models.RequestSongs:
    properties:
      hidename:
        type: integer
      id:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
...
  models.Songs:
    properties:
      from:
        type: string
      id:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
...
  models.UserRank:
    properties:
      rank:
        type: integer
    type: object
info:
  contact: {}
  description: 2020治愈系
  title: healing2020
  version: "1.0"
paths:
...
  /api/register:
    post:
      description: 注册接口
      parameters:
      - description: 用户注册数据
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
  /api/songs/rank:
    get:
      description: 每日歌曲排行榜
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AllRank'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - rank
  /api/user:
    get:
      description: 已登录用户的个人页接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PersonalPage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
    put:
      description: 更新用户信息
      parameters:
      - description: 更新的用户信息
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.PutUserInf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
...
  /api/user/rank:
    get:
      description: 用户排名
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRank'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - rank
  /api/usermodel:
    get:
      description: 获取已登录用户模型
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetUserResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
...
swagger: "2.0"